refactor(chat): move chat logic from simAction to chatMessageActions
