+++ src/actions/simAction.js
@@ -1,4 +1,4 @@
-import {createNewIMChat, parseChatFromSimulator, parseIM} from './chatMessageActions'
+import {receiveChatFromSimulator, receiveIM} from './chatMessageActions'
 import {getValueOf, mapBlockOf} from '../network/msgGetters'
 
 function parseUserRights (message) {
@@ -24,19 +24,10 @@ function simActionFilter (msg) {
 
   switch (name) {
     case 'ChatFromSimulator':
-      const parsed = parseChatFromSimulator(msg)
-      return dispatchChatAction(name, parsed, 'localchat/' + new Date(parsed.time).toJSON())
+      return receiveChatFromSimulator(msg)
 
     case 'ImprovedInstantMessage':
-      const parsedMsg = parseIM(msg)
-        // Start a new IMChat.
-      return dispatch => {
-        dispatch(createNewIMChat(
-          parsedMsg.dialog, parsedMsg.chatUUID, parsedMsg.fromId, parsedMsg.fromAgentName
-        ))
-        const id = `imChats/${parsedMsg.chatUUID}/${new Date(parsedMsg.time).toJSON()}`
-        dispatch(dispatchChatAction(name, parsedMsg, id))
-      }
+      return receiveIM(msg)
 
     case 'ChangeUserRights':
       return parseUserRights(msg)
@@ -62,34 +53,6 @@ function simActionFilter (msg) {
 // Set it in development to true to dispatch all network messages.
 if (process.env.NODE_ENV !== 'production') {
   window.debugDispatchAllMsg = window.debugDispatchAllMsg || false
-}
-
-// Dispatches chat (and IM) messages.
-// They will be saved and synced under the avatar name.
-function dispatchChatAction (name, msg, id) {
-  return async (dispatch, getState, {hoodie}) => {
-    if (typeof id !== 'string') return
-
-    const activeState = getState()
-    if (
-      activeState.account.getIn(['viewerAccount', 'loggedIn']) && activeState.account.get('sync')
-    ) {
-      // Save messages. They will also be synced!
-      msg._id = activeState.account.get('avatarIdentifier') + '/' + id
-
-      const doc = await hoodie.store.add(msg)
-      dispatch({
-        type: name,
-        msg: doc
-      })
-    } else {
-      // This is the path for every message, that will not be synced and saved.
-      dispatch({
-        type: name,
-        msg
-      })
-    }
-  }
 }
 
 function sendRegionHandshakeReply (RegionHandshake) {
